pyproject.toml
[project]
name = "code-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=0.1.0",
    "mistralai>=1.10.0",
    "openai>=2.15.0",
    "pydantic-settings>=2.7.0",
    "rich>=13.0.0",
]
